EE-248: Move tests from execution_test.rs to execution.rs
